QPPA-680: Update developer docs to use renamed fields eligiblePopulation and eligiblePopulationExclusion as well as the new field eligiblePopulationException
Closes #25

The totalPopulation field was renamed to eligiblePopulation and the
performanceExclusion field was renamed to eligiblePopulationException
and so the developer documentation needed to be updated to reflect
that. Additionally, the eligiblePopulationException field has been added
to the API so the developer documentation needed information on this new
field as well.

Reviewer: @marimiyachi