[QPPA-920]: Uses react router for routing to landing page, tutorials, developer preview and schemas.
Closes #50

https://jira.cms.gov/browse/QPPA-920

Detail:
In preparation for adding a left nav bar to all pages, I create a [tech spec](https://docs.google.com/document/d/1CL-CXHcIJifPJYUkIQQaBN15nRuDl5vSC_f3CEJv0wc/edit#) which includes a refactoring of the navigation implementation so that it is simpler to extend to additional components. This is the first PR of 3 to implement the new navigation and styling on all pages.

Test plan:
* [x] adhoc testing via UI
* [x] updated unit tests

Reviewer: @kencheeto